+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## x.y.z (not yet released)
 
   - Support instantiating arrays and collections of instantiable types from database arrays.
+  - Support binding collections as SQL arrays.
 
 ## 0.7.1 (2015-03-27)